Fix all lints

+++ pkgs/cupertinohttp/example/cupertino_client_main.dart
@@ -14,9 +14,10 @@ void main() async {
 
   final decodedResponse = jsonDecode(utf8.decode(response.bodyBytes)) as Map;
 
-  final itemCount = decodedResponse['totalItems'];
+  final itemCount = decodedResponse['totalItems'] as int;
   print('Number of books about http: $itemCount.');
   for (var i = 0; i < min(itemCount, 10); ++i) {
+    // ignore_for_file: avoid_dynamic_calls
     print(decodedResponse['items'][i]['volumeInfo']['title']);
   }
 }
